refactor: rename to topo-imagery-common
