+++ packages/topo-imagery-common/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["uv_build>=0.11.14,<0.12.0"]
 build-backend = "uv_build"
 
 [project]
-name = "common"
+name = "topo-imagery-common"
 version = "0.1.0"
 description = "Shared utilities for topo-imagery processing"
 requires-python = "~=3.12.3"
