+++ packages/topo-imagery-common/src/topo_imagery_common/aws/aws_credential_source.py
