+++ packages/topo-imagery-common/src/topo_imagery_common/files/tests/file_helper_test.py
@@ -1,5 +1,5 @@
-from common.files.files_helper import is_tiff
 from pytest_subtests import SubTests
+from topo_imagery_common.files.files_helper import is_tiff
 
 
 def test_is_tiff(subtests: SubTests) -> None:
