+++ scripts/pdal/pdal_commands.py
@@ -3,10 +3,10 @@ import subprocess
 from shutil import rmtree
 from tempfile import mkdtemp
 
-from common.aws.aws_helper import is_s3
-from common.files.fs import copy
-from common.log.time_helper import time_in_ms
 from linz_logger import get_log
+from topo_imagery_common.aws.aws_helper import is_s3
+from topo_imagery_common.files.fs import copy
+from topo_imagery_common.log.time_helper import time_in_ms
 
 
 class PDALExecutionException(Exception):
