+++ scripts/stac/imagery/capture_area.py
@@ -2,13 +2,13 @@ import json
 from decimal import Decimal
 from typing import Any, Sequence
 
-from common.log.time_helper import time_in_ms
 from linz_logger import get_log
 from shapely import BufferCapStyle, BufferJoinStyle, to_geojson, union_all, wkt
 from shapely.constructive import make_valid
 from shapely.geometry import MultiPolygon
 from shapely.geometry.base import BaseGeometry
 from shapely.ops import orient
+from topo_imagery_common.log.time_helper import time_in_ms
 
 DECIMAL_DEGREES_1M = Decimal("0.00001")
 """
