+++ scripts/stac/imagery/item.py
@@ -1,7 +1,7 @@
 import json
 from typing import Any, TypedDict
 
-from common.files.fs import read
+from topo_imagery_common.files.fs import read
 
 from scripts.stac.link import Link, Relation
 from scripts.stac.util.STAC_VERSION import STAC_VERSION
