+++ scripts/stac/imagery/tests/item_test.py
@@ -4,8 +4,8 @@ from pathlib import Path
 from typing import Any
 from unittest.mock import patch
 
-from common.files.files_helper import get_file_name_from_path
 from pytest_subtests import SubTests
+from topo_imagery_common.files.files_helper import get_file_name_from_path
 
 from scripts.conftest import any_epoch_datetime_string
 from scripts.stac.imagery.collection import ImageryCollection
