+++ scripts/stac/link.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from common.files import checksum
+from topo_imagery_common.files import checksum
 
 from scripts.stac.util.media_type import StacMediaType
