+++ scripts/standardise_validate.py
@@ -3,7 +3,8 @@ import sys
 from argparse import ArgumentParser
 from datetime import datetime, timezone
 
-from common.cli.cli_helper import (
+from linz_logger import get_log
+from topo_imagery_common.cli.cli_helper import (
     InputParameterError,
     load_input_files,
     str_to_bool,
@@ -12,10 +13,9 @@ from common.cli.cli_helper import (
     str_to_positive_int,
     valid_date,
 )
-from common.datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3339_nz_midnight_datetime_string
-from common.files.files_helper import SUFFIX_JSON, ContentType
-from common.files.fs import exists, write
-from linz_logger import get_log
+from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3339_nz_midnight_datetime_string
+from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType
+from topo_imagery_common.files.fs import exists, write
 
 from scripts.gdal.gdal_helper import get_srs, get_vfs_path
 from scripts.json_codec import dict_to_json_bytes
