+++ scripts/standardising.py
@@ -7,13 +7,13 @@ from decimal import Decimal
 from functools import partial
 from multiprocessing import Pool
 
-from common.aws.aws_helper import is_s3
-from common.cli.cli_helper import TileFiles
-from common.files.files_helper import ContentType, is_tiff
-from common.files.fs import exists, read, write, write_all, write_sidecars
-from common.log.time_helper import time_in_ms
 from linz_logger import get_log
 from tifffile import TiffFile
+from topo_imagery_common.aws.aws_helper import is_s3
+from topo_imagery_common.cli.cli_helper import TileFiles
+from topo_imagery_common.files.files_helper import ContentType, is_tiff
+from topo_imagery_common.files.fs import exists, read, write, write_all, write_sidecars
+from topo_imagery_common.log.time_helper import time_in_ms
 
 from scripts.gdal.gdal_bands import get_gdal_band_offset
 from scripts.gdal.gdal_commands import (
